refactored the tests
